Safari-611.1.10/LayoutTests/js/exception-with-handler-inside-eval-with-dynamic-scope-expected.txt

This test makes sure stack unwinding works correctly when it occurs inside an eval contained in a dynamic scope.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS result is "inner scope"
PASS result is "inner scope"
PASS result is "innermost scope"
PASS result is "innermost scope"
PASS result is "inner scope"
PASS result is "inner scope"
PASS result is "inner scope"
PASS result is "innermost scope"
PASS result is "innermost scope"
PASS result is "outer scope"
PASS result is "outer scope"
PASS successfullyParsed is true

TEST COMPLETE
